Add files via upload
Title: IRCTC Ticket Booking System

Description:

This project is a simple command-line-based implementation of a ticket booking system for the Indian Railways Catering and Tourism Corporation (IRCTC). The system offers two main functionalities: checking the train schedule and booking tickets.

**Train Schedule:**
Users can view the schedule of various trains, including details such as departure and arrival stations, departure times, train codes, and durations. The user is prompted to choose this option by entering '1'.

**Ticket Booking:**
For ticket booking, users are guided through a series of inputs to customize their booking. They can choose the train, class, source and destination stations, number of passengers, and provide passenger details like name, phone number, age, and Aadhar card number. The system calculates the fare based on the chosen parameters.

After entering the passenger details, the user is prompted to choose the mode of payment. Options include online payment or Paytm. For online payment, the user is asked to enter card details (card number, expiry date, and CVV), while for Paytm, the user needs to enter the registered phone number.

To enhance security, a captcha-like verification is implemented, where the user needs to enter a randomly generated code.

Upon successful payment, the system generates an electronic ticket (E-Ticket) with a unique PNR number, date of travel, train details, passenger information, and payment details. A trip code is also generated, which can be used to obtain seat numbers at the source station and will be sent to the registered email after the final chart.

The project is designed for educational purposes and showcases fundamental programming concepts, including user input handling, conditional statements, loops, and basic data manipulation. It provides a hands-on experience for beginners to understand the logic and flow of a simple ticket booking system.

**Note:**
This project is a text-based prototype and does not include advanced security features or user interfaces. It serves as a starting point for those interested in expanding and improving the functionality of a railway ticket booking system.